baseline implementation of RW

+++ Project/prng_engine.hpp
@@ -1,7 +1,7 @@
 //  Copyright (c) 2012-2016 M.A. (Thijs) van den Berg, http://sitmo.com/
 //
-//  Use, modification and distribution are subject to the MIT Software License. 
-//  
+//  Use, modification and distribution are subject to the MIT Software License.
+//
 //  The MIT License (MIT)
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -9,10 +9,10 @@
 //  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 //  copies of the Software, and to permit persons to whom the Software is
 //  furnished to do so, subject to the following conditions:
-//  
+//
 //  The above copyright notice and this permission notice shall be included in
 //  all copies or substantial portions of the Software.
-//  
+//
 //  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 //  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 //  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -51,7 +51,7 @@
     x1 = (x1 << rx) | (x1 >> (64-rx)); \
     z1 = (z1 << rz) | (z1 >> (64-rz)); \
     x1 ^= x0; \
-    z1 ^= z0; 
+    z1 ^= z0;
 
 
 // Double mixing function with key adition
@@ -102,7 +102,7 @@ public:
 #else
     static constexpr result_type (min)() { return 0; }
     static constexpr result_type (max)() { return 0xFFFFFFFF; }
-#endif    
+#endif
 
     // -------------------------------------------------
     // Constructors
@@ -111,7 +111,7 @@ public:
     // req: 26.5.1.4 Random number engine requirements, p.907 table 117, row 1
     // Creates an engine with the same initial state as all other
     // default-constructed engines of type E.
-    prng_engine() 
+    prng_engine()
     {
         seed();
     }
@@ -127,19 +127,19 @@ public:
         }
         _o_counter = x._o_counter;
     }
-    
-    
+
+
     // req: 26.5.1.4 Random number engine requirements, p.907 table 117, row 3
     // Creates an engine with initial O(size of state) state determined by s.
-    prng_engine(uint32_t s) 
+    prng_engine(uint32_t s)
     {
         seed(s);
     }
-    
+
         // req: 26.5.1.4 Random number engine requirements, p.908 table 117, row 4
     // Creates an engine with an initial state that depends on a sequence
     // produced by one call to q.generate.
-    template<class Seq> 
+    template<class Seq>
     prng_engine(Seq& q, typename sitmo_enable_if< has_generate_template<Seq>::value >::type* = 0 )
     {
         seed(q);
@@ -148,7 +148,7 @@ public:
     // -------------------------------------------------
     // Seeding
     // -------------------------------------------------
-    
+
     // req: 26.5.1.4 Random number engine requirements, p.908 table 117, row 5
     void seed()
     {
@@ -167,19 +167,19 @@ public:
     // req: 26.5.1.4 Random number engine requirements, p.908 table 117, row 6
     // s needs to be of return_type, which is uint32_t
     void seed(uint32_t s)
-    { 
+    {
         for (unsigned short i=0; i<4; ++i) {
             _k[i] = 0;
             _s[i] = 0;
         }
-        _k[0] = s; 
+        _k[0] = s;
         _o_counter = 0;
-        
+
         encrypt_counter();
     }
-    
+
     // req: 26.5.1.4 Random number engine requirements, p.908 table 117, row 7
-    template<class Seq> 
+    template<class Seq>
     void seed(Seq& q, typename sitmo_enable_if< has_generate_template<Seq>::value >::type* = 0 )
     {
         typename Seq::result_type w[8];
@@ -190,10 +190,10 @@ public:
             _s[i] = 0;
         }
         _o_counter = 0;
-        
+
         encrypt_counter();
     }
-    
+
     // req: 26.5.1.4 Random number engine requirements, p.908 table 117, row 8
     // Advances e’s state ei to ei+1 = TA(ei) and returns GA(ei).
     uint32_t operator()()
@@ -202,30 +202,30 @@ public:
         if (_o_counter < 8) {
             unsigned short _o_index = _o_counter >> 1;
             _o_counter++;
-            if (_o_counter&1) 
-                return _o[_o_index] & 0xFFFFFFFF; 
+            if (_o_counter&1)
+                return _o[_o_index] & 0xFFFFFFFF;
             else
                 return _o[_o_index] >> 32;
         }
-        
+
         // generate a new block and return the first 32 bits
         inc_counter();
         encrypt_counter();
         _o_counter = 1; // the next call
         return _o[0] & 0xFFFFFFFF;   // this call
     }
-    
+
     // -------------------------------------------------
     // misc
     // -------------------------------------------------
-    
+
     // req: 26.5.1.4 Random number engine requirements, p.908 table 117, row 9
     // Advances e’s state ei to ei+z by any means equivalent to z
     // consecutive calls e().
     void discard(uint64_t z)
     {
         // check if we stay in the current block
-        if (z < 8 - _o_counter) {
+        if (static_cast<int64_t>(z) < 8 - _o_counter) {
             _o_counter += static_cast<unsigned short>(z);
             return;
         }
@@ -259,12 +259,12 @@ public:
            is >> s._k[i] >> s._s[i] >> s._o[i];
        is >> s._o_counter;
        return is;
-   } 
+   }
     // req: 26.5.1.4 Random number engine requirements, p.908 table 117, row 10
-    // This operator is an equivalence relation. With Sx and Sy as the infinite 
-    // sequences of values that would be generated by repeated future calls to 
+    // This operator is an equivalence relation. With Sx and Sy as the infinite
+    // sequences of values that would be generated by repeated future calls to
     // x() and y(), respectively, returns true if Sx = Sy; else returns false.
-    bool operator==(const prng_engine& y) 
+    bool operator==(const prng_engine& y)
     {
         if (_o_counter != y._o_counter) return false;
         for (unsigned short i=0; i<4; ++i) {
@@ -274,38 +274,38 @@ public:
         }
         return true;
     }
-    
+
     // req: 26.5.1.4 Random number engine requirements, p.908 table 117, row 11
-    bool operator!=(const prng_engine& y) 
+    bool operator!=(const prng_engine& y)
     {
         return !(*this == y);
     }
-    
+
     // Extra function to set the key
     void set_key(uint64_t k0=0, uint64_t k1=0, uint64_t k2=0, uint64_t k3=0)
     {
         _k[0] = k0; _k[1] = k1; _k[2] = k2; _k[3] = k3;
         encrypt_counter();
     }
-    
+
     // set the counter
     void set_counter(uint64_t s0=0, uint64_t s1=0, uint64_t s2=0, uint64_t s3=0, unsigned short o_counter=0)
     {
-        _s[0] = s0; 
-        _s[1] = s1; 
-        _s[2] = s2; 
+        _s[0] = s0;
+        _s[1] = s1;
+        _s[2] = s2;
         _s[3] = s3;
         _o_counter = o_counter % 8;
         encrypt_counter();
     }
-    
-    
+
+
     // versioning
     uint32_t version()
     {
         return 5;
     }
-    
+
 private:
     void encrypt_counter()
     {
@@ -344,21 +344,21 @@ private:
         for (unsigned int i=0; i<4; ++i) _o[i] = b[i] + k[i];
         _o[3] += 5;
     }
-    
+
     void inc_counter()
     {
-        ++_s[0]; 
+        ++_s[0];
         if (_s[0] != 0) return;
-        
-        ++_s[1]; 
+
+        ++_s[1];
         if (_s[1] != 0) return;
-        
-        ++_s[2]; 
+
+        ++_s[2];
         if (_s[2] != 0) return;
-        
+
         ++_s[3];
     }
-    
+
     void inc_counter(uint64_t z)
     {
         if (z > 0xFFFFFFFFFFFFFFFF - _s[0]) {   // check if we will overflow the first 64 bit int
@@ -372,12 +372,12 @@ private:
         }
         _s[0] += z;
     }
-    
+
 private:
     uint64_t _k[4];             // key
     uint64_t _s[4];             // state (counter)
     uint64_t _o[4];             // cipher output    4 * 64 bit = 256 bit output
-    unsigned short _o_counter;  // output chunk counter, the 256 random bits in _o 
+    unsigned short _o_counter;  // output chunk counter, the 256 random bits in _o
                                 // are returned in eight 32 bit chunks
 };
